# backend/greenride-api-clean/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenRide API Documentation\n## Response Codes:\n- 0000: Success - 操作成功\n- 9999: System Error - 系统错误\n- 1000: Parameter Error - 参数错误\n- 1001: Authorization Error - 授权失败\n- 1002: Authentication Expired - 认证过期\n- 2001: User Error - 用户相关错误\n- 4001: Business Error - 业务错误\n- 4002: Data Expired - 数据过期",
        "title": "GreenRide API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/login": {
            "post": {
                "description": "用户登录接口，支持邮箱和手机号登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出，清除服务端相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "用户注册接口，支持乘客和司机注册。country_code为可选参数，默认为卢旺达(RW)。如果提供verify_code，会自动按优先级验证：先验证手机验证码，如果失败再验证邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或用户已存在",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/send-verify-code": {
            "post": {
                "description": "发送验证码到邮箱或手机。系统会自动根据提供的email和phone字段判断发送方式，优先使用phone。country_code为可选参数，默认为卢旺达(RW)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "发送验证码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SendVerifyCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或发送失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/verify-code": {
            "post": {
                "description": "验证邮箱或手机验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "验证验证码",
                "parameters": [
                    {
                        "description": "验证请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "user_type"
            ],
            "properties": {
                "email": {
                    "description": "login_type=email时必填",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "description": "login_type=phone时必填",
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_type": {
                    "description": "必填：passenger, driver，用于区分同一手机号/邮箱的不同身份",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                },
                "verify_code": {
                    "description": "login_type=phone时必填",
                    "type": "string"
                }
            }
        },
...
        "handlers.RegisterRequest": {
            "type": "object",
            "required": [
                "user_type",
                "username",
                "verify_code"
            ],
            "properties": {
                "country_code": {
                    "description": "可选，默认为卢旺达(RW)",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "invite_code": {
                    "description": "可选邀请码",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user_type": {
                    "description": "passenger, driver",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                },
                "username": {
                    "type": "string"
                },
                "verify_code": {
                    "description": "验证码 (可选，如果提供则优先验证手机，然后邮箱)",
                    "type": "string"
                }
            }
        },
...
        "handlers.SendVerifyCodeRequest": {
            "type": "object",
            "required": [
                "type",
                "user_type"
            ],
            "properties": {
                "country_code": {
                    "description": "手机号国家代码，默认为卢旺达(RW)",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱地址，有则发送邮箱验证码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号码，有则发送短信验证码",
                    "type": "string"
                },
                "type": {
                    "description": "验证码类型",
                    "type": "string",
                    "enum": [
                        "register",
                        "login",
                        "reset_password",
                        "change_phone",
                        "change_email"
                    ]
                },
                "user_type": {
                    "description": "用户类型，用于区分同一联系方式的不同身份",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                }
            }
        },
...
        "handlers.VerifyCodeRequest": {
            "type": "object",
            "required": [
                "code",
                "method",
                "type",
                "user_type"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "method": {
                    "description": "email, phone",
                    "type": "string"
                },
                "new_email": {
                    "description": "更换邮箱时需要",
                    "type": "string"
                },
                "new_password": {
                    "description": "重置密码时需要",
                    "type": "string"
                },
                "new_phone": {
                    "description": "更换手机时需要",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "type": {
                    "description": "验证码类型",
                    "type": "string"
                },
                "user_id": {
                    "description": "某些情况下需要用户ID",
                    "type": "string"
                },
                "user_type": {
                    "description": "用户类型，用于区分同一联系方式的不同身份",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                }
            }
        },
...
        "protocol.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}